+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 
 Also see [Issues](https://github.com/Achronite/mqtt-energenie-ener314rt/issues) for additional details.
 
-## [0.7.1] 2024-01
+## [0.7.1] 2024-02
 
 ### Added
 
@@ -31,6 +31,8 @@ Also see [Issues](https://github.com/Achronite/mqtt-energenie-ener314rt/issues) 
 * MQTT Discovery: OEM Part Number and Device ID added to HA device model field (@genestealer)
 * Bumped `energenie-ener314rt` to v0.7.1 (REQUIRED)
 * Submitting a cached command will now replace the exisiting cached command for the device
+* Added data value to the reporting of `command` when it has been succesfully cached (also only update MQTT when this is set to 0=None) [#69](https://github.com/Achronite/mqtt-energenie-ener314rt/issues/69),
+    e.g. "Set Temperature" becomes "Set Temperature 18.5"
 
 
 See also: https://github.com/Achronite/energenie-ener314rt/releases/tag/v0.7.0  - Notably pi5 support and GPIO driver changes